requirements.txt
streamlit
pdfminer.six
scikit-learn
pandas
plotly
python-dateutil
